test(#3461): move write+φ tests to dedicated file (keeps existing file untouched, all tests use canonical reason)
