--- eo-runtime/src/test/java/org/eolang/EOmallocOfPhiExpectTest.java
@@ -1,86 +1,0 @@
-/*
- * SPDX-FileCopyrightText: Copyright (c) 2016-2026 Objectionary.com
- * SPDX-License-Identifier: MIT
- */
-
-/*
- * @checkstyle PackageNameCheck (10 lines)
- * @checkstyle TrailingCommentCheck (3 lines)
- */
-package org.eolang;
-
-import org.hamcrest.MatcherAssert;
-import org.hamcrest.Matchers;
-import org.junit.jupiter.api.Assertions;
-import org.junit.jupiter.api.Test;
-
-/**
- * Test case verifying {@link Expect}-based error messages
- * raised by the {@code malloc.of.@} atom when its
- * {@code size} attribute is invalid.
- * @since 0.51
- * @checkstyle TypeNameCheck (5 lines)
- */
-@SuppressWarnings("JTCOP.RuleAllTestsHaveProductionClass")
-final class EOmallocOfPhiExpectTest {
-
-    @Test
-    void throwsCorrectErrorForNonNumericSize() {
-        MatcherAssert.assertThat(
-            "the message in the error is correct",
-            Assertions.assertThrows(
-                ExAbstract.class,
-                () -> new Dataized(
-                    new PhWith(
-                        new EOmalloc$EOof$EOφ(),
-                        Phi.RHO,
-                        new PhWith(
-                            new EOmallocOfPhiExpectTest.SizedDummy(),
-                            "size",
-                            new Data.ToPhi(true)
-                        )
-                    )
-                ).take(),
-                "malloc.of.@ with non-numeric size must fail with a proper message"
-            ).getMessage(),
-            Matchers.equalTo("the 'size' attribute must be a number")
-        );
-    }
-
-    @Test
-    void throwsCorrectErrorForNegativeSize() {
-        MatcherAssert.assertThat(
-            "the message in the error is correct",
-            Assertions.assertThrows(
-                ExAbstract.class,
-                () -> new Dataized(
-                    new PhWith(
-                        new EOmalloc$EOof$EOφ(),
-                        Phi.RHO,
-                        new PhWith(
-                            new EOmallocOfPhiExpectTest.SizedDummy(),
-                            "size",
-                            new Data.ToPhi(-1)
-                        )
-                    )
-                ).take(),
-                "malloc.of.@ with negative size must fail with a proper message"
-            ).getMessage(),
-            Matchers.equalTo("the 'size' attribute (-1) must be greater or equal to zero")
-        );
-    }
-
-    /**
-     * Minimal Phi with a single {@code size} attribute, used as a
-     * stand-in for the {@code malloc.of} parent in tests where {@code size}
-     * validation must fail before the {@code allocated}/{@code scope}
-     * attributes are touched.
-     * @since 0.51
-     */
-    private static final class SizedDummy extends PhDefault {
-
-        SizedDummy() {
-            super(new Attrs(new AttrEntry("size", new AtVoid("size"))));
-        }
-    }
-}
